when2meet/pyproject.toml
[project]
name = "when2meet"
version = "0.1.0"
description = "When2Meet clone project backend"
authors = [
    {name = "LuSiEDa", email = "wltlrtmdgus@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "orjson (>=3.11.2,<4.0.0)",
    "tortoise-orm[asyncmy] (>=0.25.1,<0.26.0)",
    "cryptography (>=45.0.6,<46.0.0)",
    "aerich (>=0.9.1,<0.10.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "24.10.0"
ruff = "0.8.2"
mypy = "^1.17.1"
pytest = "^8.4.1"
coverage = "^7.10.4"
pytest-asyncio = "^1.1.0"

[tool.black]
line-length = 120

[tool.ruff]
target-version = "py313"

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = 3.13
strict = true
explicit_package_bases = true
mypy_path = "when2meet"

[tool.coverage.run]
omit = ["*/test_*.py"]  # 루트 혹은 루트 아래의 파일들 중 test_ 로 시작하는 파일을 제외.

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
filterwarnings = ['ignore::DeprecationWarning']

[tool.aerich]
tortoise_orm = "app.configs.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

